r/build_config: making the `step` block Optional+Computed
When creating a Build Configuration based on a Template, once the
build has first been created - Terraform shows a diff because the
Build Steps are missing from the configuration but returned from
the API.

As such this commit makes this block `Optional` & `Computed` so that
this diff can be omitted if a user hasn't specified any `step` blocks
to override this in the config (e.g. wants to /only/ use steps from
the template).